+++ common/src/main/kotlin/com/rpmtw/rpmtw_platform_mod/translation/machineTranslation/QueueText.kt
@@ -1,5 +1,5 @@
 package com.rpmtw.rpmtw_platform_mod.translation.machineTranslation
 
-import java.util.*
+import com.rpmtw.rpmtw_platform_mod.translation.language.TranslateLanguage
 
-data class QueueText(val text: String, val language: Locale)
+data class QueueText(val text: String, val language: TranslateLanguage)
